requirements.txt
google-cloud-bigquery==3.0.1
google-cloud-storage==2.3.0
pandas==1.3.5
db-dtypes==1.0.0
tqdm==4.64.0
joblib==1.1.0
pyarrow==7.0.0
gcsfs==2022.5.0
google-api-core==2.7.2